inline transfer rpc handlers, remove no-op waitUntil calls
